% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constant_metadata.R
\name{add_constant_metadata}
\alias{add_constant_metadata}
\title{add_constant_metadata}
\usage{
add_constant_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{list: List returned by fin_to_ddh_keys(), used in add_new_datasets()}
}
\value{
list
}
\description{
Add metadata that have constant values across records (not pulled from the Finance API)
}
