% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constant_metadata_resource.R
\name{add_constant_metadata_resource}
\alias{add_constant_metadata_resource}
\title{add_constant_metadata_resource}
\usage{
add_constant_metadata_resource(metadata_list)
}
\arguments{
\item{metadata_list}{list: List for machine names and their corresponding values}
}
\value{
list
}
\description{
Add metadata that have constant values across records for resources
}
