% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_dataset.R
\name{add_new_dataset}
\alias{add_new_dataset}
\title{add_new_dataset}
\usage{
add_new_dataset(metadata_list, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{metadata_list}{list: list with one finance dataset's metadata, from get_fin_datasets_metadata()}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the dkanr::get_credentials() function}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
