% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_datasets.R
\name{add_new_datasets}
\alias{add_new_datasets}
\title{add_new_datasets}
\usage{
add_new_datasets(metadata_lists, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{metadata_lists}{list: list of finance metadata, from get_fin_datasets_metadata()}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the dkanr::get_credentials() function}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
