% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_resource_to_dataset.R
\name{attach_resource_to_dataset}
\alias{attach_resource_to_dataset}
\title{attach_resource_to_dataset
link resources to dataset}
\usage{
attach_resource_to_dataset(dataset_nid, body, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{dataset_nid}{character: Node ID of the dataset to be updated}

\item{body}{json: json object generated by jsonlite::toJSON}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the get_credentials() function}
}
\description{
attach_resource_to_dataset
link resources to dataset
}
