% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_dataset.R
\name{create_json_dataset}
\alias{create_json_dataset}
\title{create_json_dataset}
\usage{
create_json_dataset(metadata_list,
  json_template = fin2ddh::ddh_schema_finance_dataset)
}
\arguments{
\item{metadata_list}{list: List of metatadata values}

\item{json_template}{list: List generated from JSON template}
}
\value{
json
}
\description{
Create a json according to predefined template
}
