% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_resource.R
\name{create_json_resource}
\alias{create_json_resource}
\title{create_json_resource}
\usage{
create_json_resource(metadata_list,
  json_template = mdlibtoddh::json_template_resource)
}
\arguments{
\item{metadata_list}{list: List of metatadata values}

\item{json_template}{list: List generated from JSON template}
}
\value{
json
}
\description{
Create a json according to predefined template
}
