% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fin_metadata.R
\name{filter_fin_metadata}
\alias{filter_fin_metadata}
\title{filter_fin_metadata}
\usage{
filter_fin_metadata(finance_data)
}
\arguments{
\item{finance_data}{list: The formatted JSON from extract_fin_metadata}
}
\value{
list
}
\description{
Filter metadata that is in catalog
}
