% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_to_ddh_keys.R
\name{fin_to_ddh_keys}
\alias{fin_to_ddh_keys}
\title{fin_to_ddh_keys
Extract specific metadata from the Finance API JSON response}
\usage{
fin_to_ddh_keys(metadata_in, metadata_out = mdlibtoddh::md_placeholder,
  lookup = fin2ddh::lookup)
}
\arguments{
\item{metadata_in}{list: The output of get_fin_datasets_metadata()}

\item{metadata_out}{list: Package object: mdlibtoddh::md_placeholder}

\item{lookup}{data.frame: Package object: fin2ddh::lookup}
}
\value{
list
}
\description{
fin_to_ddh_keys
Extract specific metadata from the Finance API JSON response
}
