% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metadata.R
\name{format_metadata}
\alias{format_metadata}
\title{format_metadata}
\usage{
format_metadata(metadata_in,
  timestamp_fields = c("field_wbddh_modified_date",
  "field_wbddh_release_date"), date_fields = c("field_wbddh_start_date",
  "field_wbddh"))
}
\arguments{
\item{metadata_in}{list: flattened list of metadata}

\item{date_fields}{character: variables to be processed as dates}

\item{text_fields}{character: variables to be processed as text}
}
\value{
list
}
\description{
Format metadata fields to the format accepted by ddh
}
