% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fin_datasets_metadata.R
\name{get_fin_datasets_metadata}
\alias{get_fin_datasets_metadata}
\title{get_fin_datasets_metadata}
\usage{
get_fin_datasets_metadata(fin_internal_ids = NULL,
  is_unit_test = FALSE)
}
\arguments{
\item{fin_internal_ids}{vector: list of finanace ids to subset the number of datasets added}

\item{is_unit_test}{boolean: check if unit test is being run}
}
\value{
list
}
\description{
Get all finance metadata for those datasets in the finance catalog
}
