% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fin_metadata.R
\name{map_fin_metadata}
\alias{map_fin_metadata}
\title{map_fin_metadata.R}
\usage{
map_fin_metadata(metadata_list)
}
\arguments{
\item{metadata}{list:}
}
\value{
list
}
\description{
Extract specific metadata from the Microdata API JSON response
}
