% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_date_read.R
\name{recursive_date_read}
\alias{recursive_date_read}
\title{read values from the date fields which require a recursive read}
\usage{
recursive_date_read(metadata_in)
}
\arguments{
\item{metadata_in}{list: a dataset extracted from the finances data portal}
}
\value{
list
}
\description{
read values from the date fields which require a recursive read
}
