% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_created_dataset.R
\name{test_created_dataset}
\alias{test_created_dataset}
\title{test_created_dataset}
\usage{
test_created_dataset(dataset_nid, metadata_list, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{dataset_nid}{character: The dataset node id}

\item{metadata_list}{list: List of metatadata values}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
character
}
\description{
Test the metadata values of the dataset tranferred to DDH
}
