% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_to_ddhdate.R
\name{timestamp_to_ddhdate}
\alias{timestamp_to_ddhdate}
\title{timestamp_to_ddhdate}
\usage{
timestamp_to_ddhdate(time_stamp, origin = "1970-01-01")
}
\arguments{
\item{time_stamp}{numeric: A linux time stamp}

\item{origin}{character the date of origin}
}
\value{
character vector
}
\description{
timestamp_to_ddhdate
}
\examples{
timestamp_to_ddhdate(time_stamp = 1497978729, origin = "1970-01-01")
}
