% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublish_existing_dataset.R
\name{unpublish_existing_dataset}
\alias{unpublish_existing_dataset}
\title{unpublish_existing_dataset}
\usage{
unpublish_existing_dataset(dataset_nid, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()),
  root_url = dkanr::get_url())
}
\arguments{
\item{credentials}{list: DDH API authentication token and cookie}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}

\item{metadata_list}{list: Finance portal internal ID of the dataset to be added}

\item{master}{dataframe: Output of fin2ddh::get_ddh_records_status()}
}
\value{
character
}
\description{
Update a full finance record in DDH (metadata + resources)
}
