% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fin_metadata.R
\docType{data}
\name{url}
\alias{url}
\title{extract_fin_metadata}
\format{An object of class \code{character} of length 1.}
\usage{
url
}
\arguments{
\item{url}{string: Financa data api}
}
\value{
list
}
\description{
Extract specific metadata from the Microdata API JSON response
}
\keyword{datasets}
