% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_link_to_resources.R
\name{add_link_to_resources}
\alias{add_link_to_resources}
\title{add_link_to_resources}
\usage{
add_link_to_resources(metadata_list, category)
}
\arguments{
\item{metadata_list}{list: Flattened list of metadata from the Fianace API}

\item{category}{string: finance category retrieved from metadata$view$category}
}
\value{
list
}
\description{
add_link_to_resources
}
