% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_dataset.R
\name{add_new_dataset}
\alias{add_new_dataset}
\title{add_new_dataset}
\usage{
add_new_dataset(url)
}
\arguments{
\item{url}{string: Financa data api}
}
\value{
list
}
\description{
Extract specific metadata from the Microdata API JSON response
}
