% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_dataset.R
\name{add_new_dataset}
\alias{add_new_dataset}
\title{add_new_dataset}
\usage{
add_new_dataset(root_url = dkanr::get_url(), ddh_credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{url}{string: Financa data api}
}
\value{
list
}
\description{
Extract specific metadata from the Microdata API JSON response
}
