% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_datasets.R
\name{add_new_datasets}
\alias{add_new_datasets}
\title{add_new_datasets}
\usage{
add_new_datasets(metadata_lists, root_url = dkanr::get_url(),
  ddh_credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{metadata_lists}{list: list of finance metadata}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
