% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_unpublish.R
\name{create_json_unpublish}
\alias{create_json_unpublish}
\title{create_json_unpublish}
\usage{
create_json_unpublish()
}
\value{
json
}
\description{
Create a json according to predefined template to unpublish dataset
}
