% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fin_metadata.R
\name{extract_fin_metadata}
\alias{extract_fin_metadata}
\title{extract_fin_metadata}
\usage{
extract_fin_metadata(is_unit_test = FALSE)
}
\arguments{
\item{is_unit_test}{boolean: check if unit test is being run}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
