% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fin_datasets_list.R
\name{get_fin_datasets_list}
\alias{get_fin_datasets_list}
\title{get_fin_datasets_list
return all the finance datasets with updated time}
\usage{
get_fin_datasets_list(is_unit_test = FALSE)
}
\arguments{
\item{is_unit_test}{boolean: check if unit test is being run}
}
\value{
data.frame
}
\description{
get_fin_datasets_list
return all the finance datasets with updated time
}
