% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_created_dataset.R
\name{test_created_dataset}
\alias{test_created_dataset}
\title{test_created_dataset}
\usage{
test_created_dataset(dataset_metadata, metadata_list,
  lovs = ddhconnect::get_lovs(), root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{dataset_metadata}{character: The dataset node id}

\item{metadata_list}{list: List of metatadata values}

\item{lovs}{dataframe: lookup table of data catalog machine names and tids, return object of get_lovs()}

\item{root_url}{character: API root URL}

\item{credentials}{list: list with dkanr::get_cookie() and dkanr::get_token()}
}
\value{
character
}
\description{
Test the metadata values of the dataset tranferred to DDH
}
