% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_created_dataset.R
\name{test_created_dataset}
\alias{test_created_dataset}
\title{test_created_dataset}
\usage{
test_created_dataset(nid, metadata_list, credentials,
  root_url = production_root_url)
}
\arguments{
\item{nid}{character: The dataset node id}

\item{metadata_list}{list: List of metatadata values}

\item{credentials}{list: object returned by the get_credentials() function}

\item{root_url}{character: API root URL}
}
\value{
character
}
\description{
Test the metadata values of the dataset tranferred to DDH
}
