% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublish_existing_dataset.R
\name{unpublish_existing_dataset}
\alias{unpublish_existing_dataset}
\title{unpublish_existing_dataset}
\usage{
unpublish_existing_dataset(dataset_nid, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{dataset_nid}{string: Finance dataset that needs to be unpublished}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}

\item{credentials}{list: DDH API authentication token and cookie}
}
\value{
character
}
\description{
Update a full finance record in DDH (metadata + resources)
}
