% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_dataset.R
\name{update_existing_dataset}
\alias{update_existing_dataset}
\title{update_existing_dataset}
\usage{
update_existing_dataset(metadata_lists,
  master = fin2ddh::get_ddh_records_status(), root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{metadata_lists}{list: list of finance metadata, from get_fin_datasets_metadata()}

\item{master}{dataframe: Output of fin2ddh::get_ddh_records_status()}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the dkanr::get_credentials() function}
}
\value{
character
}
\description{
Update a full finance record in DDH (metadata + resources)
}
