% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_dataset.R
\name{update_existing_dataset}
\alias{update_existing_dataset}
\title{update_existing_dataset()}
\usage{
update_existing_dataset(fin_internal_id, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()), master,
  root_url = dkanr::get_url())
}
\arguments{
\item{fin_internal_id}{character: Finance portal internal ID of the dataset to be added}

\item{credentials}{list: DDH API authentication token and cookie}

\item{master}{dataframe: Master lookup table}

\item{root_url}{character: Root URL to use for the API (Staging or Production)}
}
\value{
character
}
\description{
Update a full finance record in DDH (metadata + resources)
}
